requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
dnspython==1.16.0
easyocr==1.7.2
filelock==3.19.1
Flask==3.1.2
fsspec==2025.9.0
google==3.0.0
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pyclipper==1.3.0.post6
pycparser==2.23
pymongo==3.12.0
python-bidi==0.6.6
python-dotenv==1.1.1
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.2
soupsieve==2.8
sympy==1.14.0
tifffile==2025.9.30
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
Werkzeug==3.1.3
